refactor: remove ValueWrapper
