+++ core/Cargo.toml
@@ -36,5 +36,7 @@ humantime-serde = "1.0.1"
 bloomfilter = "1.0.5"
 dynfmt = { version = "0.1.5", features = [ "curly" ] }
 sqlx = { version = "0.5.7", features = [ "postgres", "mysql", "runtime-async-std-native-tls", "decimal", "chrono" ] }
+async-std = "1.6.4"
 uriparse = "0.6.3"
 paste = "1.0"
+rust_decimal = "1.10.3"
